+++ src/get_links3.py
@@ -1,6 +1,5 @@
 # Code written by:
-# Joshua Loh
-# Denton Phosavanh
+# Joshua Loh, Denton Phosavanh
 # 2016
 
 # Webscraping
@@ -10,6 +9,7 @@ from bs4 import SoupStrainer
 
 # Other
 import argparse
+import textwrap
 import sys
 from queue import *
 
@@ -19,70 +19,76 @@ def streamline(a):
 	a.sort()
 	return a
 
+
 # From a string in the form of "Wikipedia_Article_Name",
 # create the BeautifulSoup object of it, and return
 # a list of all the links that it contains
 def get_links(s):
 	# Set up the scraper
 	wiki = "https://en.wikipedia.org/wiki/" + s
-	page = urllib.request.urlopen(wiki)
-	# soup = BeautifulSoup(page, "html.parser")
-	soup = BeautifulSoup(page, "html.parser", parse_only=SoupStrainer("p"))
+	
+	try:
+		page = urllib.request.urlopen(wiki)
+	except urllib.error.HTTPError:
+		print("An error occurred in opening this page")
+		print("This is most likely due to either of two reasons:")
+		print("a) Your internet connection isn't internetting")
+		print("b) The wikipedia page %s does not exist" % wiki)
+		exit(3)
+	except Exception:
+		import traceback
+		exit(4)
+	else:
+		# only load the parts of the page contained in a <p>
+		soup = BeautifulSoup(page, "html.parser", parse_only=SoupStrainer("p"))
 
 	# Stores the name of all the outgoing Wiki pages
 	link_array = []
 
 	# Extract every <a> that is in a <p>, store them in link_array
-	# every_para = soup.find_all("p")
-	# for para in every_para:
-	# 	# print para.get_text()
 	para_links = soup.find_all("a")
 	for para_link in para_links:
 		link = para_link.get("href")
 		# Only pull (relevant parts of) relevant pages
-		if link.startswith("/wiki/") and (link[6:].startswith("Wikipedia:") == False) and (link[6:].startswith("Help:") == False):
+		if link.startswith("/wiki/") and (not link[6:].startswith("Wikipedia:")) and (not link[6:].startswith("Help:")):
 			link_array.append(link[6:])
 
 	# Return link_array
 	return streamline(link_array)
 
+
 # Recursive DFS
-def dfs(s, depth, DEPTH_LIMIT):
-	# Base cases, have reached depth limit, or found Hitler
+def dfs(site, depth, DEPTH_LIMIT, target, verbosity):
+	# Base cases, have reached depth limit, or found target
 	if depth >= DEPTH_LIMIT:
 		return
-		# exit(11)
-	# elif s == "Adolf_Hitler" or s == "Hitler":
+	
+	# elif s == target:
 	# 	print("YEAH YEAH YEAH YEAH YEAH YEAH YEAH")
 	# 	print("Got to " + s + " at depth " + str(depth))
 	# 	exit(12)
 	global visited_links
-	visited_links.add(s)
+	visited_links.add(site)
 	indent = "\t"*depth
-	print(indent + "Visiting: " + s)
+	print(indent + "Visiting: " + site)
 
 	# Output all the links that can be seen from this page
 	indent += "\t"
-	link_array = get_links(s)
+	link_array = get_links(site)
 	for l in link_array:
 		print(indent + l)
 
-	# Shortcut search for Hitler
-	# for l in link_array:
-	# 	if l == "Adolf_Hitler" or l == "Hitler":
-	# 		dfs(l, depth+1, DEPTH_LIMIT)
-
-	# recursion
-	# Another sneaky base case
+	# recursion (Another sneaky base case)
 	for l in link_array:
 		if not l in visited_links:
-			dfs(l, depth+1, DEPTH_LIMIT)
+			dfs(l, depth+1, DEPTH_LIMIT, target, verbosity)
 
-def bfs(s):
+
+def bfs(site, target, verbosity):
 	global visited_links
 	fringe = Queue()
-	fringe.put(s)
-	print(s)
+	fringe.put(site)
+	print(site)
 	# Keep searching until have run out of potential expansions
 	while (not fringe.empty()):
 		current_link = fringe.get()
@@ -91,7 +97,7 @@ def bfs(s):
 		for l in link_array:
 			print("Child: " + l)
 			# Check for exit state
-			if l == "Aboriginal_peoples_in_Canada" or l == "Hitler":
+			if l == "Aboriginal_peoples_in_Canada" or l == target:
 				print("Found")
 				exit(13)
 
@@ -102,14 +108,39 @@ def bfs(s):
 
 
 def argparse_setup():
-	parser = argparse.ArgumentParser()
+	parser = argparse.ArgumentParser(
+		formatter_class=argparse.RawDescriptionHelpFormatter,
+		# Describe the program and arguments briefly
+		description=textwrap.dedent("""\
+			Finds a path between pages [start] and [target] on Wikipedia
+			Set options on how the program will run
+			Default values:
+			---------------
+			 a : bfs
+			 d : 2
+			 s : Sun_Dance
+			 t : Adolf_Hitler
+			 v : 1
+			---------------
+			"""),
+			
+		# Gloss over exit details
+		epilog=textwrap.dedent("""\
+			Exit status:
+			------------
+			 0 : Everything worked as planned
+			 2 : Error parsing arguments
+			 3 : urllib.error.HTTPError
+			 4 : General exception from urllib.request
+			""")
+	)
 	
 	# Which search to use
 	parser.add_argument(
-		"-s", "--search",
+		"-a", "--algorithm",
 		default = "bfs",
 		choices = ["dfs", "bfs"],
-		help = "which search to use (default: bfs)"	
+		help = "which search to use"	
 	)
 	
 	# dfs depth
@@ -117,13 +148,21 @@ def argparse_setup():
 		"-d", "--depth",
 		default = 2,
 		type = int,
-		help = "the depth to which dfs should search (default = 2). use in conjunction with dfs)"
+		help = "maximum search depth of dfs. use in conjunction with dfs)"
 	)
 	
+	# Starting page
 	parser.add_argument(
-		"-p", "--page",
+		"-s", "--start",
 		default = "Sun_Dance",
 		help = "which page to start the search from (formatted to Wikipedia standards)"
+	)
+	
+	# Target page
+	parser.add_argument(
+		"-t", "--target",
+		default = "Adolf_Hitler",
+		help = "the page for which to search"
 	)
 	
 	# Level of output
@@ -131,34 +170,27 @@ def argparse_setup():
 		"-v", "--verbosity",
 		default = 1,
 		type = int,
-		choices = [0, 1, 2, 3, 4, 5],
-		help = "the level of output to use (0 is low, 5 is high)"
+		help = "the level of output to use (higher => more output)"
 	)
 	
 	return parser
-	
+
+
 #+-------------------------------
 #| main
 #+-------------------------------
-# If no argument is passed, exit
-# if len(sys.argv) < 3:
-	# print("Error, invalid number of arguments.")
-	# print("Example usage: " + sys.argv[0] + " [Article_Name] [DFS_DEPTH]")
-	# sys.exit()
-
+# Parse the program's arguments
 parser = argparse_setup()
 args = parser.parse_args()
-print(args)
+if args.verbosity >= 1:
+	print(args)
 
 # Keep track of which links have been visited
 visited_links = set()
-visited_links.add(args.page)
+visited_links.add(args.start)
 
-if args.search == "dfs":
-	dfs(args.page, 0, args.depth)
+# Run either dfs or bfs, depending on arguments
+if args.algorithm == "dfs":
+	dfs(args.start, 0, args.depth, args.target, args.verbosity)
 else:
-	bfs(args.page)
-	
-# visited_links.add(sys.argv[1])
-# dfs(sys.argv[1], 0, int(sys.argv[2]))
-# bfs(sys.argv[1])
+	bfs(args.start, args.target, args.verbosity)